[add] Abstraction layer on API branches to get rid of globals and straighten execution flow
