+++ speculos/__init__.py
@@ -1,4 +1,3 @@
 from . import api  # noqa: F401
 from . import client  # noqa: F401
 from . import mcu  # noqa: F401
-from . import main  # noqa: F401
