+++ speculos/api/__init__.py
@@ -1,0 +1,2 @@
+from .api import ApiRunner # noqa: F401
+from .events import events # noqa: F401
