+++ speculos/api/swagger.py
@@ -1,0 +1,6 @@
+from .restful import AppResource
+
+
+class Swagger(AppResource):
+    def get(self):
+        return self.app.send_static_file("swagger/index.html")
