+++ speculos/api/web_interface.py
@@ -1,0 +1,6 @@
+from .restful import AppResource
+
+
+class WebInterface(AppResource):
+    def get(self):
+        return self.app.send_static_file("index.html")
