+++ speculos/main.py
@@ -19,7 +19,7 @@ import threading
 
 import pkg_resources
 
-from .api import api
+from .api import ApiRunner, events
 from .mcu import apdu as apdu_server
 from .mcu import automation
 from .mcu import display
@@ -28,6 +28,7 @@ from .mcu.automation_server import AutomationClient, AutomationServer
 from .mcu.button_tcp import FakeButton
 from .mcu.finger_tcp import FakeFinger
 from .mcu.vnc import VNC
+
 
 DEFAULT_SEED = ('glory promote mansion idle axis finger extra february uncover one trip resource lawn turtle enact '
                 'monster seven myth punch hobby comfort wild raise skin')
@@ -269,14 +270,14 @@ def main(prog=None):
     if args.automation_port:
         logger.warn("--automation-port is deprecated, please use the REST API instead")
         if api_enabled:
-            logger.warn("--automation-port is incompatible with the the API server, disabling it")
+            logger.warn("--automation-port is incompatible with the the API server, disabling the latter")
             api_enabled = False
         automation_server = AutomationServer(("0.0.0.0", args.automation_port), AutomationClient)
         automation_thread = threading.Thread(target=automation_server.serve_forever, daemon=True)
         automation_thread.start()
 
     if api_enabled:
-        automation_server = api.events
+        automation_server = events
 
     s1, s2 = socket.socketpair()
 
@@ -317,7 +318,7 @@ def main(prog=None):
 
     apirun = None
     if api_enabled:
-        apirun = api.ApiRunner(args.api_port)
+        apirun = ApiRunner(args.api_port)
 
     display_args = display.DisplayArgs(args.color, args.model, args.ontop, rendering, args.keymap, zoom, x, y)
     server_args = display.ServerArgs(apdu, apirun, button, finger, seph, vnc)
